breaking: change callOperation to accept a request object
This is a breaking change as the paramters of the method
BaseService::callOperation have changed. Code calling this method must
now pass an instance of the BaseType class as the second parameter. The
method will now construct the XML request within itself by calling the
BaseType::toRequestXml method on the passed request object.