feature: allow attachments to be sent and received
The SDK now allows attachments to be sent as part of the request.
Likewise attachments are handled if they appear in the response.

To add an attachment to the request object simply call the `attachment`
method passing in the binary data of the attachment as the first
parameter. Note that you do not have to base64 encode the data!

```php
$request->attachment(file_get_contents(__DIR__.'/picture.jpg'));
```

To get the attachment from a response simply call the same method with
no parmaters. The method will return an associative array with two keys.
The key 'data' is the binary data of the attachment while the key
'mimeType' returns the mime type.

```php
$response = $service->downloadFile($request);

$attachment = $response->attachment();

$fp = fopen('attachment', 'wb');
fwrite($fp, $attachment['data']);
fclose($fp);
```